Problem_A/requirements.txt
pdfplumber>=0.11.0
pypdf2>=3.0.0
anthropic>=0.84.0
python-docx>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
